requirements.txt
asyncpg
openai
pydantic-ai
logfire
PyMuPDF
pypdf
tqdm
python-dotenv
tiktoken
